2.9.30

+++ fluent-crm/app/Http/Controllers/CampaignController.php
@@ -36,19 +36,27 @@ class CampaignController extends Controller
         $order = $request->get('sort_type') ?: 'DESC';
         $orderBy = $request->get('sort_by') ?: 'id';
         $with = $request->get('with', []);
+        $labels = $request->getSafe('labels', [], 'intval');
 
-        $campaigns = Campaign::when($status, function ($query) use ($status) {
+        $campaignQuery = Campaign::when($status, function ($query) use ($status) {
             return $query->whereIn('status', $status);
         })->when($search, function ($query) use ($search) {
             return $query->where('title', 'LIKE', "%$search%");
         })
-            ->orderBy($orderBy, ($order == 'ASC') ? 'ASC' : 'DESC')
-            ->paginate();
+            ->orderBy($orderBy, ($order == 'ASC') ? 'ASC' : 'DESC');
 
+        if (!empty($labels)) {
+            $campaignQuery->whereHas('labelsTerm', function ($query) use ($labels) {
+                $query->whereIn('term_id', $labels);
+            });
+        }
+
+        $campaigns = $campaignQuery->paginate();
         if (in_array('stats', $with)) {
             foreach ($campaigns as $campaign) {
                 $campaign->stats = $campaign->stats();
                 $campaign->next_step = fluentcrm_get_campaign_meta($campaign->id, '_next_config_step', true);
+                $campaign->labels = $campaign->getFormattedLabels();
             }
         }
 
@@ -90,6 +98,15 @@ class CampaignController extends Controller
             $campaign = Campaign::findOrFail($id);
         }
 
+        /**
+         * Determine the email campaign data in FluentCRM.
+         *
+         * This filter allows modification of the email campaign data before it is used.
+         *
+         * @since 2.6.51
+         *
+         * @param array $campaign The email campaign data.
+         */
         $campaign = apply_filters('fluent_crm/campaign_data', $campaign);
 
         $templates = Template::emailTemplates()
@@ -305,6 +322,16 @@ class CampaignController extends Controller
         do_action('fluentcrm_campaign_status_active', $campaign);
 
         $page = (int)$this->request->get('page', 1);
+        /**
+         * Determine the number of subscribers to process per request in FluentCRM Email Campaign.
+         *
+         * This filter allows you to modify the number of subscribers that are processed
+         * in a single request when handling email campaigns.
+         *
+         * @since 2.7.0
+         * 
+         * @param int The number of subscribers to process per request. Default is 90.
+         */
         $limit = (int)apply_filters('fluent_crm/process_subscribers_per_request', 90);
 
         $subscribersSettings = [
@@ -685,9 +712,53 @@ class CampaignController extends Controller
         $preHeader = (!empty($campaign->email_pre_header)) ? $campaign->email_pre_header : '';
 
         if ($subscriber) {
+            /**
+             * Determine the email campaign body text before it is sent.
+             *
+             * This filter allows you to modify the email body content for a campaign before it is sent to the subscriber.
+             *
+             * @since 2.7.0
+             * 
+             * @param string $emailBody The email body content.
+             * @param object $subscriber The subscriber object.
+             *
+             * @return string The filtered email body content.
+             */
             $emailBody = apply_filters('fluent_crm/parse_campaign_email_text', $emailBody, $subscriber);
+            /**
+             * Determine the email footer text for a campaign.
+             *
+             * This filter allows you to modify the email footer text for a campaign before it is sent to a subscriber.
+             *
+             * @since 2.7.0 
+             * 
+             * @param string $emailFooter The email footer text.
+             * @param object $subscriber The subscriber object.
+             */
             $emailFooter = apply_filters('fluent_crm/parse_campaign_email_text', $emailFooter, $subscriber);
+            /**
+             * Determine the email campaign subject text.
+             *
+             * This filter allows you to modify the email subject text for a campaign.
+             *
+             * @since 2.7.0
+             * 
+             * @param string $emailSubject The email subject text.
+             * @param object $subscriber The subscriber object.
+             */
             $emailSubject = apply_filters('fluent_crm/parse_campaign_email_text', $emailSubject, $subscriber);
+            /**
+             * Determine the pre-header text of the campaign email.
+             *
+             * This filter allows you to modify the pre-header text of the campaign email before it is sent to the subscriber.
+             *
+             * @since 2.7.0
+             * 
+             * @param string $preHeader The pre-header text of the campaign email.
+             * @param object $subscriber The subscriber object containing subscriber details.
+             *
+             * @return string The filtered pre-header text.
+             */
             $preHeader = apply_filters('fluent_crm/parse_campaign_email_text', $preHeader, $subscriber);
         }
 
@@ -698,6 +769,18 @@ class CampaignController extends Controller
             'config'      => wp_parse_args(Arr::get($campaign->settings, 'template_config', []), Helper::getTemplateConfig($campaign->design_template))
         ];
 
+        /**
+         * Determine the email body content based on the design template type.
+         *
+         * This filter allows modification of the email body content based on the specified design template.
+         *
+         * @since 2.5.1
+         * 
+         * @param string $emailBody The email body content.
+         * @param array $templateData The data used for the email template.
+         * @param object $campaign The campaign object.
+         * @param object $subscriber The subscriber object.
+         */
         $emailBody = apply_filters(
             'fluent_crm/email-design-template-' . $campaign->design_template,
             $emailBody,
@@ -786,7 +869,31 @@ class CampaignController extends Controller
         $emailFooter = Helper::getEmailFooterContent($campaign);
 
         if ($subscriber) {
+            /**
+             * Determine the campaign email body content text.
+             *
+             * This filter allows you to modify the email body content before it is sent to the subscriber.
+             *
+             * @since 2.7.0
+             * 
+             * @param string $emailBody The original email body content.
+             * @param object $subscriber The subscriber object containing subscriber details.
+             *
+             * @return string The filtered email body content.
+             */
             $emailBody = apply_filters('fluent_crm/parse_campaign_email_text', $emailBody, $subscriber);
+            /**
+             * Determine the campaign email footer text.
+             *
+             * This filter allows you to modify the email footer content before it is sent to the subscriber.
+             *
+             * @since 2.7.0
+             * 
+             * @param string $emailFooter The original email footer content.
+             * @param object $subscriber The subscriber object containing subscriber details.
+             *
+             * @return string The filtered email footer content.
+             */
             $emailFooter = apply_filters('fluent_crm/parse_campaign_email_text', $emailFooter, $subscriber);
         }
 
@@ -797,6 +904,16 @@ class CampaignController extends Controller
         $preHeader = (!empty($campaign->email_pre_header)) ? $campaign->email_pre_header : '';
 
         if ($preHeader && $subscriber) {
+            /**
+             * Determine the campaign email Pre-header text before sending.
+             *
+             * This filter allows you to modify the email pre-header content for a campaign before it is sent to the subscriber.
+             *
+             * @since 2.7.0
+             *
+             * @param string $emailBody The email body content.
+             * @param object $subscriber The subscriber object.
+             */
             $preHeader = apply_filters('fluent_crm/parse_campaign_email_text', $preHeader, $subscriber);
         }
 
@@ -807,6 +924,18 @@ class CampaignController extends Controller
             'config'      => wp_parse_args(Arr::get($campaign->settings, 'template_config', []), Helper::getTemplateConfig($campaign->design_template))
         ];
 
+        /**
+         * Determine the email body content based on the design template type.
+         *
+         * This filter allows modification of the email body content based on the specified design template.
+         *
+         * @since 2.5.1
+         * 
+         * @param string $emailBody The email body content.
+         * @param array $templateData The data used for the email template.
+         * @param object $campaign The campaign object.
+         * @param object $subscriber The subscriber object.
+         */
         $emailBody = apply_filters(
             'fluent_crm/email-design-template-' . $designTemplate,
             $emailBody,
@@ -818,7 +947,18 @@ class CampaignController extends Controller
         if (strpos($emailBody, '{{crm') || strpos($emailBody, '##crm')) {
             $emailBody = str_replace(['{{crm_global_email_footer}}', '{{crm_preheader_text}}'], [$emailFooter, $preHeader], $emailBody);
             if (strpos($emailBody, '##crm.') || strpos($emailBody, '{{crm.')) {
-                // we have CRM specific smartcodes
+                /**
+                 * Determine the email body content including a specific Smartcode for a subscriber.
+                 *
+                 * This filter allows modification of the email body content including a smartcode before it is sent to a subscriber.
+                 *
+                 * @since 2.7.0
+                 * 
+                 * @param string $emailBody The email body content.
+                 * @param object $subscriber The subscriber object.
+                 *
+                 * @return string The filtered email body content.
+                 */
                 $emailBody = apply_filters('fluent_crm/parse_extended_crm_text', $emailBody, $subscriber);
             }
         }
@@ -856,6 +996,7 @@ class CampaignController extends Controller
 
     public function handleBulkAction(Request $request)
     {
+        $actionName = $request->getSafe('action_name', '');
         $campaignIds = $request->get('campaign_ids', []);
         $campaignIds = array_map(function ($id) {
             return (int)$id;
@@ -869,18 +1010,44 @@ class CampaignController extends Controller
             ]);
         }
 
-        $campaigns = Campaign::whereIn('id', $campaignIds)->get();
-        foreach ($campaigns as $campaign) {
-            $campaignId = $campaign->id;
-            $campaign->deleteCampaignData();
-            $campaign->delete();
-            do_action('fluent_crm/campaign_deleted', $campaignId);
+        if ($actionName == 'apply_labels') {
+            $newLabels = $request->getSafe('labels', '');
+
+            if (!$newLabels) {
+                return $this->sendError([
+                    'message' => __('Please provide labels', 'fluent-crm')
+                ]);
+            }
+
+            $campaigns = Campaign::whereIn('id', $campaignIds)->get();
+
+            foreach ($campaigns as $campaign) {
+                $campaign->attachLabels($newLabels);
+            }
+
+            return $this->sendSuccess([
+                'message' => __('Labels has been applied successfully', 'fluent-crm'),
+            ]);
+
         }
 
-        return $this->sendSuccess([
-            'message' => __('Selected Campaigns has been deleted permanently', 'fluent-crm'),
-        ]);
+        if ($actionName == 'delete_campaigns') {
+            $campaigns = Campaign::whereIn('id', $campaignIds)->get();
+            foreach ($campaigns as $campaign) {
+                $campaignId = $campaign->id;
+                $campaign->deleteCampaignData();
+                $campaign->delete();
+                do_action('fluent_crm/campaign_deleted', $campaignId);
+            }
 
+            return $this->sendSuccess([
+                'message' => __('Selected Campaigns has been deleted permanently', 'fluent-crm'),
+            ]);
+        }
+
+        return $this->sendError([
+            'message' => __('invalid bulk action', 'fluent-crm')
+        ]);
     }
 
     public function createTemplate()
@@ -949,7 +1116,7 @@ class CampaignController extends Controller
                 if (($requestCounter % 4) === 0) {
                     $lastChecked = fluentCrmGetOptionCache('_fcrm_last_email_process_cleanup', 600);
                     if (!$lastChecked || time() - $lastChecked > 140) {
-                        $dateStamp = date('Y-m-d H:i:s', (current_time('timestamp') - 150));
+                        $dateStamp = gmdate('Y-m-d H:i:s', (current_time('timestamp') - 150));
                         CampaignEmail::where('status', 'processing')
                             ->where('updated_at', '<', $dateStamp)
                             ->update([
@@ -1013,7 +1180,7 @@ class CampaignController extends Controller
                 $maximumProcessingTime = fluentCrmMaxRunTime() + 40;
                 CampaignEmail::where('campaign_id', $campaignId)
                     ->where('status', 'processing')
-                    ->where('updated_at', '<', date('Y-m-d H:i:s', (current_time('timestamp') - $maximumProcessingTime)))
+                    ->where('updated_at', '<', gmdate('Y-m-d H:i:s', (current_time('timestamp') - $maximumProcessingTime)))
                     ->update([
                         'status' => 'pending'
                     ]);
@@ -1031,7 +1198,6 @@ class CampaignController extends Controller
                     $campaign = Campaign::withoutGlobalScope('type')->findOrFail($campaignId);
 
                     do_action('fluent_crm/campaign_archived', $campaign);
-
                 }
             }
         }
@@ -1185,8 +1351,10 @@ class CampaignController extends Controller
             'created_by'       => get_current_user_id(),
             'settings'         => $oldCampaign->settings
         ];
+        $labelIds = $oldCampaign->getFormattedLabels()->pluck('id')->toArray();
 
         $campaign = Campaign::create($newCampaign);
+        $campaign->attachLabels($labelIds);
 
         $campaign->duplicateSubjects($oldCampaign);
 
@@ -1248,5 +1416,25 @@ class CampaignController extends Controller
         return [
             'sharable_url' => $campaign->getShareableUrl()
         ];
+    }
+
+    public function updateLabels(Request $request, $funnel_id)
+    {
+        $funnel = Campaign::findOrFail($funnel_id);
+        $action = $request->getSafe('action');
+        $labelIds = $request->getSafe('label_ids');
+
+        if (!is_array($labelIds)) {
+            $labelIds = [$labelIds];
+        }
+
+        if ($action == 'detach') {
+            $funnel->detachLabels($labelIds);
+        }
+
+        return $this->sendSuccess([
+            'message' => __('Labels has been updated', 'fluent-crm')
+        ]);
+
     }
 }
